refactor: extract capture/evidence/ocr routes into routes/captures.ts
Pure structural move out of createApp; no handler logic changed. Moves
POST /captures, GET /captures/recent, GET /cases/:id/captures/count,
GET /cases/:id/evidence/:file, and GET /cases/:id/ocr-search verbatim.

Graduated onto RouteContext (shared with the drop-watch ingest path and the
AI-control routes still in createApp): getControl, flush, indexCaptureText.
Capture-only state (lastCapture, the SIGNIFICANT trigger set, windowSize,
evidenceContentType) moves into the module; buffers is reused via the existing
ctx.captureBuffers() live accessor.
