+++ companion/src/routes/context.ts
@@ -3,6 +3,7 @@ import type { CaseStore } from "../storage/caseStore.js";
 import type { Logger } from "../logging/logger.js";
 import type { AppOptions } from "../server.js";
 import type { CaptureMetadata } from "../types.js";
+import type { AiControl } from "../analysis/aiControl.js";
 import type { ImporterRegistry } from "../analysis/importerStore.js";
 import type { IrisClient } from "../integrations/iris/irisClient.js";
 import type { EnrichmentProvider } from "../enrichment/provider.js";
@@ -35,6 +36,16 @@ export interface RouteContext {
   recordAiError(caseId: string, phase: string, err: unknown): void;
   readUnlockState(req: Request, id: string, salt: string): { unlocked: boolean; remembered: boolean };
   hasAiProvider(): boolean;
+  // Capture→analyze machinery shared with the drop-watch ingest path and the AI-control routes
+  // (all still in createApp). Graduated for routes/captures.ts's POST /captures handler:
+  //   getControl         — read the per-case AI on/off + last-analyzed-seq control record.
+  //   flush              — drain a case's capture buffer through the analysis pipeline.
+  //   indexCaptureText   — queue a persisted screenshot for background OCR full-text indexing.
+  // Stable (hoisted function declarations bound at construction); the LIVE state they touch
+  // (buffers, synth in-flight, the OCR queue) is reached through the live accessors below.
+  getControl(caseId: string): Promise<AiControl>;
+  flush(caseId: string): Promise<void>;
+  indexCaptureText(metadata: CaptureMetadata): void;
 
   // ── LIVE accessors ───────────────────────────────────────────────────────────────────
   // Call these INSIDE the request handler (or inside per-request logic like a preflight run),
